start of a bash sudoku script
Includes an initial object model that I will probably change to be
something efficient for lookups. Entries are currently encoded as
strings "ijn" where 'i' is the row, 'j' is the column, and 'n' is the
value at that coordinate. To look for entries at a coordinate (i,j),
one needs to inspect entries in 'startingBoard' and 'guesses' arrays
until either finding a match or else all entries have been checked.

Includes an initial printBoard function that has coloration
  red    for startingBoard values
  yellow for guesses
  green  for the board
but it doesn't print out the full board boundries lines yet.