Add validation rows, columns, and subsquares; whole board
Added validation functions for each part of the board, and incorporated
them into the whole-board validation function.

There is some logging that needs polishing. Right now it reports missing
entries in parts of the board, but it short-circuits in tandem with how
the whole-board completion check short-circuits. Instead of printing out
missing entries for every row, column, and subsquare, it prints out
missing values for the /first/ row with missing values (since rows are
validated first). If all values show up in all rows, it moves on to do
the same thing on columns (since columns are validated second). If all
values also show up in all columns, it moves on to do the same thing on
subsquares (since subsquares are validated last).