Board completion validation is no longer short-circuited
This will be useful for a pending update where we will add flagging and
extra formatting for entries that are "invalid" (i.e., entries that are
duplicated in their row, column, or subsquare). We must examine/validate
all rows, columns, and subsquares to be able to flag all such entries.