Add formatting (highlighting) for invalid/duplicate entries
Use a new Bash array to track when an entry needs invalid / (duplicated
in its row/column/subsquare). Extra formatting is currently in the form
of highlighting (background color), and the bg color is customized
depending on whether the entry is a guess or is from the starting board.

I have some echoDebug calls saved in this commit. Will remove them in
the next commit.

I start using the "local" keyword for some function variables in this
commit. This is good practice that I should use more consistently, to
avoid accidentally borrowing a value from the wrong context, and to be
more explicit about which vars are local and which are global.